finished up and tested all the crud operations

+++ src/main/java/com/example/springboottraining/service/ToDoService.java
@@ -12,6 +12,6 @@ public interface ToDoService {
     List<ToDo> getAllToDo();
     List<ToDo> getFinishedToDo();
     List<ToDo> getOpenToDo();
-    Integer getNumOfOpenToDo();
-    Integer getNumOfFinishedToDo();
+    Long getNumOfOpenToDo();
+    Long getNumOfFinishedToDo();
 }
